src/requirements.txt
langchain==0.0.352
langchain_experimental==0.0.47
langsmith==0.0.74
pandas==1.4.2
streamlit==1.29.0
streamlit_feedback==0.1.3
openai==0.28.1
neo4j==5.14.1

psycopg2==2.9.3
python-environ==0.4.54
python-dotenv==1.0.0

argilla=1.21.0
